requirements.txt
spacy
regex
nltk
spellchecker
tensorflow
keras
protobuf
tensorflow-datasets
tensorflow-estimator
transformers
bertviz
numpy
pandas
faiss-cpu
openai
langchain
en-core-web-sm
matplotlib
langchain-community
langchain-core